+++ packages/components/src/Button/ButtonTransparent.tsx
@@ -25,39 +25,14 @@
  */
 
 import styled from 'styled-components'
-import { variant } from 'styled-system'
 import { ButtonBase } from './ButtonBase'
 
 /* eslint-disable sort-keys-fix/sort-keys-fix */
-const transparentButtonPadding = variant({
-  prop: 'size',
-  variants: {
-    xxsmall: {
-      px: 'xsmall',
-    },
-    xsmall: {
-      px: 'xsmall',
-    },
-    small: {
-      px: 'small',
-    },
-    medium: {
-      px: 'large',
-    },
-    large: {
-      px: 'large',
-    },
-  },
-})
-
 export const ButtonTransparent = styled(ButtonBase)`
   background: transparent;
   border: 1px solid transparent;
   color: ${({ theme, color = 'key' }) => theme.colors[color]};
-  ${(props) =>
-    props.iconBefore || props.iconAfter
-      ? transparentButtonPadding
-      : `padding: 0 ${props.theme.space.xsmall};`}
+  padding: 0 ${(props) => props.theme.space.xsmall};
 
   &:hover,
   &:focus,